fashionclip/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
transformers>=4.46
torch>=2.6
pillow>=11
